create db for favorite recipe

+++ app/src/main/java/com/allana/food_recipe_app_chapter7/data/local/room/datasource/FavoriteRecipeDataSource.kt
@@ -3,7 +3,7 @@ package com.allana.food_recipe_app_chapter7.data.local.room.datasource
 import com.allana.food_recipe_app_chapter7.data.local.room.entity.FavoriteRecipe
 
 interface FavoriteRecipeDataSource {
-    suspend fun insertFavoriteRecipe(favRecipe: FavoriteRecipe): Int
+    suspend fun insertFavoriteRecipe(favRecipe: FavoriteRecipe): Long
     suspend fun deleteFavoriteRecipe(favRecipe: FavoriteRecipe): Int
 
 }
